rightBox내부 비디오 부분 컴포넌트화 및 VideoList 레이아웃구성
